version 1.2

+++ uninstall.php
@@ -1,5 +1,8 @@
 <?php
-if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+if (!defined('ABSPATH')) {
+    exit;
+}
+if (!defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit();
 }
 delete_option('quick-ajax-global-options');
